Maintenance: Tagged 1.8.4 release for VMS after Raymond successfully              built and tested distribution on the VMS server.

+++ test/dt_arith.c
@@ -5218,7 +5218,7 @@ main(void)
     unsigned long	nerrors = 0;
 
     /* Set the random # seed */
-    HDsrandom((unsigned)HDtime(NULL));
+    HDsrandom((unsigned long)HDtime(NULL));
 
     reset_hdf5();
